migrate dummy to tests
