+++ burn-compute/src/tests/integration/mod.rs
@@ -1,0 +1,2 @@
+mod dummy;
+mod dummy_server;
