+++ burn-compute/src/tests/mod.rs
@@ -1,0 +1,3 @@
+mod integration;
+mod memory_management;
+mod storage;
